feat(app): create searchContainer own component
