+++ src/pages/ProductSearch/index.scss
@@ -44,44 +44,3 @@
     width: 90%;
   }
 }
-
-/**homepage styles**/
-
-.homepage {
-  margin: 1rem;
-
-  &__heading {
-    margin-block: 1rem;
-    font-weight: 700;
-    text-align: center;
-  }
-
-  &__search {
-    display: flex;
-    justify-content: center;
-
-    &-container {
-      display: flex;
-      flex-direction: column;
-      gap: 0.25rem;
-    }
-
-    &-button {
-      > span {
-        display: none;
-      }
-    }
-  }
-}
-
-@media only screen and (min-width: 992px) {
-  .homepage {
-    &__search {
-      &-button {
-        > span {
-          display: initial;
-        }
-      }
-    }
-  }
-}
